requirements.txt
python-binance
numpy
talib